requirements.txt
accelerate==1.10.1
aiocache==0.12.3
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.14.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.8.0
APScheduler==3.10.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.8.1
asitop==0.0.24
async-timeout==5.0.1
attrs==25.1.0
Authlib==1.6.3
av==16.0.1
azure-ai-documentintelligence==1.0.2
azure-core==1.36.0
azure-identity==1.25.0
azure-storage-blob==12.24.1
babel==2.17.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.12.3
bidict==0.23.1
black==25.9.0
blessed==1.20.0
blinker==1.9.0
boto3==1.40.5
botocore==1.40.52
Brotli==1.1.0
build==1.2.2.post1
cachetools==5.5.1
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==1.0.20
click==8.1.8
cohere==5.13.11
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
cryptography==44.0.0
cssselect2==0.7.0
csvw==3.5.1
ctranslate2==4.6.0
dashing==0.1.0
dataclasses-json==0.6.7
ddgs==9.0.0
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
dlinfo==2.0.0
docstring_parser==0.17.0
docx2txt==0.8
duckduckgo_search==7.3.0
durationpy==0.9
ecdsa==0.19.1
einops==0.8.1
emoji==2.15.0
espeakng-loader==0.2.4
et_xmlfile==2.0.0
Events==0.5
faiss-cpu==1.10.0
fake-useragent==2.2.0
fastapi==0.118.0
fastavro==1.10.0
faster-whisper==1.1.1
feedparser==6.0.12
ffmpy==0.5.0
filelock==3.17.0
filetype==1.2.0
flatbuffers==25.1.24
fonttools==4.60.1
fpdf2==2.8.2
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.2.3
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.38.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gradio==5.14.0
gradio_client==1.7.0
greenlet==3.1.1
gritql==0.1.5
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.2
huggingface-hub==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kokoro-onnx==0.4.6
kubernetes==32.0.0
lance-namespace==0.4.5
lance-namespace-urllib3-client==0.4.5
lancedb==0.29.1
langchain==0.3.27
langchain-chroma==0.2.1
langchain-cli==0.0.35
langchain-community==0.3.29
langchain-core==0.3.79
langchain-experimental==0.3.4
langchain-huggingface==0.1.2
langchain-ollama==0.2.3
langchain-text-splitters==0.3.11
langdetect==1.0.9
langserve==0.3.1
langsmith==0.4.35
language-tags==1.2.0
ldap3==2.9.1
litellm==1.59.10
lmstudio==1.4.1
loguru==0.7.3
lxml==5.3.0
Mako==1.3.10
Markdown==3.9
markdown-it-py==3.0.0
markdownify==0.14.1
MarkupSafe==2.1.5
marshmallow==3.26.0
mcp==1.14.1
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
msal==1.34.0
msal-extensions==1.3.1
msgspec==0.19.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.24.1
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
ollama==0.4.7
onnxruntime==1.20.1
open-webui==0.6.33
open_text_embeddings==1.0.4
openai==1.60.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opensearch-py==2.8.0
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pdfminer.six==20231228
pdfplumber==0.11.5
peewee==3.18.1
peewee-migrate==1.12.2
phonemizer-fork==3.3.1
pillow==11.3.0
platformdirs==4.5.0
posthog==3.11.0
primp==0.15.0
prompt-template==1.1.0
propcache==0.2.1
proto-plus==1.26.1
protobuf==5.29.3
psutil==7.0.0
pyarrow==20.0.0
pyasn1==0.4.8
pyasn1_modules==0.4.1
pybase64==1.4.2
pyclipper==1.3.0.post6
pycparser==2.22
pycrdt==0.12.25
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pymdown-extensions==10.14.2
PyMySQL==1.1.1
pypandoc==1.15
pyparsing==3.2.3
pypdf==6.0.0
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.12.3
python-iso639==2025.2.18
python-jose==3.4.0
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
python-socketio==5.13.0
pytokens==0.1.10
pytube==15.0.0
pytz==2025.1
pyxlsb==1.0.10
PyYAML==6.0.2
rank-bm25==0.2.2
RapidFuzz==3.14.1
rapidocr-onnxruntime==1.4.4
rdflib==7.1.4
redis==7.0.0b3
referencing==0.36.2
regex==2023.12.25
requests==2.32.5
requests-mock==1.12.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
RestrictedPython==8.0
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
ruff==0.9.4
s3transfer==0.13.1
safehttpx==0.1.6
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
segments==2.3.0
semantic-chunkers==0.0.10
semantic-router==0.0.68
semantic-version==2.10.0
sentence-transformers==5.1.1
sentencepiece==0.2.1
setuptools==75.8.0
sgmllib3k==1.0.0
shapely==2.1.2
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
smolagents==1.6.0
sniffio==1.3.1
socksio==1.0.0
soundfile==0.13.1
soupsieve==2.6
SQLAlchemy==2.0.38
sse-starlette==1.8.2
starlette==0.45.3
starlette-compress==1.6.0
starsessions==2.2.1
streamlit==1.41.1
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.4.0
tokenizers==0.21.0
toml==0.10.2
tomlkit==0.13.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.48.2
typer==0.9.4
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
unstructured==0.16.17
unstructured-client==0.42.3
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.37.0
uvloop==0.21.0
validators==0.35.0
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
wrapt==1.17.2
wsproto==1.2.0
xgboost==3.1.2
xlrd==2.0.1
xlsxwriter==3.2.9
yarl==1.18.3
youtube-transcript-api==1.2.2
zipp==3.21.0
zstandard==0.23.0
